Added finish, colors for start and finish, RTS-style camera movement
For start and finish added StartScript and FinishScript which basically only change the colors of the tiles under start and finish flags. Also give coordinates of start and finish to MapManager.
All camera movement script will be stored in CameraMovement.cs which is attached to Main Camera in the scene.