Added play button, PriorityQueue class, and implemented Dijkstra
Still working through a bug where i (a numerable in percDown and startAlgorithm) extends past the bounds of the list. Most likely is just a problem with the way I wrote PriorityQueue or Dijkstra. Also fixed a bug with the camera where it would continue moving post user holding down right-click.