Fixed bug with Dijkstra and neighbors, found another
Neighbors weren't getting properly found (the top row and final column outside of the tilemap were thought to be there by the system even when they weren't) and fixing that fixed the indexing outside of the tilemap error. There are now two prominent bugs: you must hit the play button TWICE to start the Dijkstra properly, the first results in a false error. The second bug is the heapList of Tuples aren't automatically updating their distance which makes Dijkstra randomly pick its next node, instead of the true next minChild.